# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AttachmentResponse:
    properties:
      duration_ms:
        type: integer
      file_name:
        type: string
      height:
        type: integer
      mime_type:
        type: string
      size:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
...
  dto.ChatResponse:
    properties:
      id:
        type: integer
      is_read_only:
        type: boolean
      last_message_text:
        type: string
      meetup:
        $ref: '#/definitions/dto.MeetupResponse'
      meetup_id:
        type: integer
      title:
        type: string
      type:
        type: string
      unread_count:
        type: integer
    type: object
  dto.Coordinates:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
...
  dto.MeetupResponse:
    properties:
      address:
        type: string
      coordinates:
        $ref: '#/definitions/dto.Coordinates'
      cover_url:
        type: string
      creator:
        $ref: '#/definitions/dto.ProfileResponse'
      creator_id:
        type: integer
      description:
        type: string
      distance_meters:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      invite_token:
        type: string
      is_member:
        type: boolean
      is_public:
        type: boolean
      participants:
        items:
          $ref: '#/definitions/dto.ProfileResponse'
        type: array
      participants_count:
        type: integer
      start_time:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.TagResponse'
        type: array
      title:
        type: string
    type: object
  dto.MessagePreview:
    properties:
      content:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      sender_display_name:
        type: string
      sender_id:
        type: integer
    type: object
  dto.MessageResponse:
    properties:
      attachment:
        $ref: '#/definitions/dto.AttachmentResponse'
      chat_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      edited_at:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      message_type:
        type: string
      reply_to:
        $ref: '#/definitions/dto.MessagePreview'
      request_id:
        type: string
      sender:
        $ref: '#/definitions/dto.ProfileResponse'
      sender_id:
        type: integer
    type: object
  dto.PresenceResponse:
    properties:
      display_name:
        type: string
      last_seen:
        type: integer
      online:
        type: boolean
      user_id:
        type: integer
    type: object
  dto.ProfileResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_organizer:
        type: boolean
      tags:
        items:
          $ref: '#/definitions/dto.TagResponse'
        type: array
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  dto.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  handler.EditMessageRequest:
    properties:
      content:
        example: Исправленный текст
        type: string
    type: object
  handler.SendMessageRequest:
    properties:
      content:
        example: Привет, как дела?
        type: string
      file_id:
        example: 6f5e4d3c-2b1a-...
        type: string
      reply_to_message_id:
        example: 42
        type: integer
      request_id:
        description: |-
          RequestID — клиентский idempotency key (UUID). Повторный POST с тем же
          request_id не создаёт дубль, а возвращает уже созданное сообщение (200).
        example: a1b2c3d4-...
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: API Server for Mobile Application "Meetuper"
  title: Offline Meetup API
  version: "1.0"
paths:
...
  /v1/chats:
    get:
      description: Возвращает список всех чатов (групповых и личных), в которых состоит
        текущий авторизованный пользователь. Отсортированы по активности (чаты с новыми
        сообщениями — сверху). Включает текст последнего сообщения и счетчик непрочитанных.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ChatResponse'
            type: array
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список чатов пользователя
      tags:
      - Chats
  /v1/chats/{id}/messages:
    get:
      description: Возвращает историю сообщений для конкретного чата. Поддерживает
        cursor-based пагинацию. Пользователь должен быть участником чата.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: ID последнего полученного сообщения (для пагинации, по умолчанию
          0)
        in: query
        name: cursor
        type: integer
      - description: Лимит сообщений (по умолчанию 50, максимум 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MessageResponse'
            type: array
        "400":
          description: Неверный ID чата или параметры
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение сообщений чата
      tags:
      - Chats
    post:
      consumes:
      - application/json
      description: Сохраняет новое сообщение в базу и моментально рассылает его онлайн-участникам
        чата через WebSockets.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Текст сообщения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отправка сообщения
      tags:
      - Chats
  /v1/chats/{id}/messages/{messageId}:
    delete:
      description: Удаляет своё сообщение (мягкое удаление). Удалять может только
        автор. Удаление рассылается участникам чата через WebSocket.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: ID сообщения
        in: path
        name: messageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Удалено
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Не автор сообщения
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Сообщение не найдено
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление сообщения
      tags:
      - Chats
    patch:
      consumes:
      - application/json
      description: Меняет текст своего сообщения. Редактировать может только автор.
        Изменение рассылается участникам чата через WebSocket.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: ID сообщения
        in: path
        name: messageId
        required: true
        type: integer
      - description: Новый текст
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.EditMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Не автор сообщения
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Сообщение не найдено
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Редактирование сообщения
      tags:
      - Chats
  /v1/chats/{id}/presence:
    get:
      description: Возвращает онлайн-статус и время последнего визита для всех участников
        чата. Вызывающий должен быть участником.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PresenceResponse'
            type: array
        "400":
          description: Неверный ID чата
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Нет доступа к чату
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Присутствие участников чата
      tags:
      - Chats
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"